Automatically push backups to GitHub.

+++ packages/backups.js
@@ -9,7 +9,16 @@ exports.SaveBackup = async function() {
 			return void console.log(stderr)
 		console.log("Backup completed!")
 		const backups = fs.readdirSync("backups").length
-		if (backups > 10)
-			exec("cd backups && rm -r \"$(ls -t | tail -1)\"")
+
+		if (backups > 10) exec("cd backups && rm -r \"$(ls -t | tail -1)\"")
+
+		// Update backups automatically.
+		exec('git add backups', (error, stdout, stderr) => {
+			if (error) return void console.log(stderr);
+			exec('git commit -m "Update backups (AUTOMATED BY BB)"', (error, stdout, stderr) => {
+				if (error) return void console.log(stderr);
+				exec("git push");
+			})
+		})
 	})
 }
